refactor: userLike 테이블 user-service로 이동후 post-service 로직 수정중

+++ post-service/src/main/java/com/sobok/postservice/post/dto/response/PostDetailResDto.java
@@ -19,7 +19,7 @@ public class PostDetailResDto {
     private String cookName;
     private String nickname;
     private Long userId;
-    private int likeCount;
+    private Long likeCount;
     private List<String> images;
     private Long updatedAt;
     private String content;
